changed totalvalue from integer to double because money is not always an integer

+++ src/eldar/creditcard/models/CreditCard.java
@@ -9,14 +9,14 @@ public class CreditCard {
     private final Integer number;
     private final Person holder;
     private final LocalDate expireDate;
-    private final Integer totalSpent;
+    private final Double totalSpent;
 
     public CreditCard(CardIssuer issuer, Integer number, Person holder, LocalDate expireDate) {
         this.issuer = issuer;
         this.number = number;
         this.holder = holder;
         this.expireDate = expireDate;
-        this.totalSpent = 0;
+        this.totalSpent = 0.0;
     }
 
     public CardIssuer getIssuer() {
@@ -35,7 +35,7 @@ public class CreditCard {
         return expireDate;
     }
 
-    public Integer getTotalSpent() {
+    public Double getTotalSpent() {
         return totalSpent;
     }
